requirements.txt
# Core frameworks
google-adk
streamlit>=1.30.0

# Google & Gemini APIs
google-generativeai
google-cloud-aiplatform

# Data & Map
pandas
folium
streamlit-folium

# Utilities
python-dotenv